Re-work initialisation of stack.
`gcc` doesn't like the way the pointers to the stack are being casted,
so instead, get the linker script to manufacture some symbols and use
those in our code.

Icky point: as the descriptor table is a table of function pointers, the
stack has to be a "function pointer" too or `gcc` will complain about an
invalid data type.